integreer softdeletes

+++ database/migrations/2017_09_28_221825_create_registration_table.php
@@ -19,7 +19,7 @@ class CreateRegistrationTable extends Migration
             $table->string('lastname');
             $table->integer('age');
             $table->string('adress');
-            $table->tinyInteger('housenumber');
+            $table->Integer('housenumber');
             $table->string('municipality');
             $table->integer('postalcode');
             $table->string('email', 100)->unique();
